ModuleRobot is done
